# target/generated-sources/openapi/src/main/resources/openapi.yaml
openapi: 3.0.2
info:
  contact:
    email: info@inss.ch
  description: |-
    Some sample APIs collected by International Network Support & Service - Glas.
    This document contains endpoints containing APIs that don't have only common formats like application/json, but more uncommon but standard formats like float numbers and special characters like german Umlaute.
    This is to test & check if a tool can handle also those formats.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://inss.ch
  title: "Petstore, the troublemaker of OpenAPI 3.0"
  version: 1.0.1
...
servers:
- url: http://localhost:8080/api/v3
...
paths:
...
  /oops:
    get:
      description: Produces sample error response.
      operationId: failingRequest
      responses:
        "200":
          description: Never returned.
        "304":
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              explode: false
              schema:
                type: string
              style: simple
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestError'
          description: Bad request.
      summary: Always fails with 400.
      tags:
      - troublemaker
      x-accepts: application/json
      x-tags:
      - tag: troublemaker
      - tag: failing
...
  /pet/{petId}:
    delete:
      description: Deletes a pet.
      operationId: deletePet
      parameters:
      - description: Api key.
        explode: false
        in: header
        name: API_KEY
        required: false
        schema:
          example: apikeyvalue1
          type: string
        style: simple
      - description: Pet id to delete
        explode: false
        in: path
        name: petId
        required: true
        schema:
          example: 99
          format: int64
          type: integer
        style: simple
      responses:
        "400":
          description: Invalid pet value
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: Deletes a pet
      tags:
      - pet
      x-accepts: application/json
      x-tags:
      - tag: pet
    get:
      description: Returns a single pet with example of a german Umlaut.
      operationId: getPetById
      parameters:
      - description: ID of pet to return
        explode: false
        in: path
        name: petId
        required: true
        schema:
          example: 99
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Pet not found
      security:
      - api_key: []
      - petstore_auth:
        - write:pets
        - read:pets
      summary: Sample data contains german umlaut.
      tags:
      - pet
      x-accepts: application/json
      x-tags:
      - tag: pet
      - tag: troublemaker
    post:
      description: Updates a pet in the store with form data.
      operationId: updatePetWithForm
      parameters:
      - description: ID of pet that needs to be updated
        explode: false
        in: path
        name: petId
        required: true
        schema:
          example: 99
          format: int64
          type: integer
        style: simple
      - description: Name of pet that needs to be updated
        explode: true
        in: query
        name: name
        required: false
        schema:
          example: dog
          type: string
        style: form
      - description: Status of pet that needs to be updated
        explode: true
        in: query
        name: status
        required: false
        schema:
          example: in progress.
          type: string
        style: form
      responses:
        "405":
          description: Invalid input
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: Updates a pet in the store with form data
      tags:
      - pet
      x-accepts: application/json
      x-tags:
      - tag: pet
  /pet/uploadbinplusobject:
    post:
      description: Uploads a binary file including a json object.
      operationId: uploadBinaryPlusObject
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/uploadBinaryPlusObject_request'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelApiResponse'
          description: successful operation
        "500":
          description: internal error.
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: uploads a binary file plus json
      tags:
      - petobject
      x-content-type: multipart/form-data
      x-accepts: application/json
      x-tags:
      - tag: petobject
      - tag: troublemaker
  /pet/uploadbinary:
    post:
      description: Uploads a binary file.
      operationId: uploadBinary
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/uploadBinary_request'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelApiResponse'
          description: successful operation
        "500":
          description: internal error.
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: uploads a binary file
      tags:
      - petobject
      x-content-type: multipart/form-data
      x-accepts: application/json
      x-tags:
      - tag: petobject
...
components:
...
  schemas:
...
    Category:
      example:
        name: Dogs
        id: 1
      properties:
        id:
          description: Id
          example: 1
          format: int64
          type: integer
        name:
          description: Name of category.
          example: Dogs
          type: string
      type: object
...
    Tag:
      example:
        name: tag1
        id: 1
      properties:
        id:
          description: Id
          example: 1
          format: int64
          type: integer
        name:
          description: Name of tag.
          example: tag1
          type: string
      type: object
    Pet:
      example:
        photoUrls: "[http://example.com/photo.png, http://example.com/photo2,png]"
        name: SChäferhund
        id: 10
        category:
          name: Dogs
          id: 1
        tags:
        - name: tag1
          id: 1
        - name: tag1
          id: 1
        status: available
      properties:
        id:
          description: Id
          example: 10
          format: int64
          type: integer
        name:
          description: Name
          example: SChäferhund
          type: string
        category:
          $ref: '#/components/schemas/Category'
        photoUrls:
          description: Photo urls.
          example: "[http://example.com/photo.png, http://example.com/photo2,png]"
          items:
            type: string
          type: array
        tags:
          items:
            $ref: '#/components/schemas/Tag'
          type: array
        status:
          description: pet status in the store
          enum:
          - available
          - pending
          - sold
          example: available
          type: string
      required:
      - name
      - photoUrls
      type: object
    RestError:
      description: The schema for all error responses.
      properties:
        status:
          description: The HTTP status code.
          example: 400
          format: int32
          readOnly: true
          title: Status
          type: integer
        error:
          description: The short error message.
          example: Bad Request
          readOnly: true
          title: Error
          type: string
        path:
          description: The path of the URL for this request.
          example: /api/owners
          format: uri
          readOnly: true
          title: Path
          type: string
        timestamp:
          description: The time the error occured.
          format: date-time
          readOnly: true
          title: Timestamp
          type: string
        message:
          description: The long error message.
          example: Request failed schema validation
          readOnly: true
          title: Message
          type: string
        trace:
          description: The stacktrace for this error.
          example: "com.atlassian.oai.validator.springmvc.InvalidRequestException:\
            \ ..."
          readOnly: true
          title: Trace
          type: string
      required:
      - error
      - message
      - path
      - schemaValidationErrors
      - status
      - timestamp
      title: REST Error
      type: object
    ModelApiResponse:
      example:
        code: 123
        type: responsetype
        message: Message.
      properties:
        code:
          description: Code
          example: 123
          format: int32
          type: integer
        type:
          description: Response type
          example: responsetype
          type: string
        message:
          description: The message.
          example: Message.
          type: string
      type: object
    uploadBinaryPlusObject_request:
      properties:
        pet:
          $ref: '#/components/schemas/Pet'
        fileName:
          format: binary
          type: string
      required:
      - pet
      type: object
    uploadBinary_request:
      properties:
        orderId:
          example: 999
          format: int32
          type: integer
        userId:
          example: 10001000
          format: int64
          type: integer
        fileName:
          format: binary
          type: string
      type: object
  securitySchemes:
    petstore_auth:
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
      type: oauth2
    api_key:
      in: header
      name: api_key
      type: apiKey